requirements.txt
apturl==0.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.6.2
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.6
colorama==0.4.3
command-not-found==0.3
coverage==7.5.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.11
django-debug-toolbar==4.3.0
django-nose==1.4.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.1
Faker==24.14.1
fasteners==0.14.1
Flask==2.3.2
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
monotonic==1.5
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
packaging==24.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.5.0
protobuf==3.6.1
psycopg2==2.8.6
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pytest==8.2.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
typing-extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
waitress==2.1.2
Werkzeug==2.3.6
xkit==0.0.0
zipp==3.16.2
